refactor cachedClient into client package and add ssaPatch
